move tsconfig to devtools
